create components character and mapping data strong in every character
